feat(REFIT-24809): four decimals

+++ source/src/main/java/com/sixgroup/refit/observability/item35/creator/domain/model/Storage.java
@@ -3,14 +3,16 @@ package com.sixgroup.refit.observability.item35.creator.domain.model;
 import lombok.Data;
 import lombok.RequiredArgsConstructor;
 
+import java.math.BigDecimal;
+
 @RequiredArgsConstructor
 @Data
 public class Storage {
 
     private String timeStamp;
-    private Float capacity;
+    private BigDecimal capacity;
 
-    public Storage(String timeStamp, Float capacity) {
+    public Storage(String timeStamp, BigDecimal capacity) {
         this.timeStamp = timeStamp;
         this.capacity = capacity;
     }
